argparse -> click; bio_hansel -> biohansel; creating base for `biohansel create`
