+++ biohansel/subtype/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""
+"""
+
+from biohansel.subtype.subtype import Subtype
+from biohansel.subtype.subtyper import subtype_contigs_samples, subtype_reads_samples
