+++ biohansel/subtype/const.py
@@ -1,17 +1,16 @@
 # -*- coding: utf-8 -*-
 
-import re
 from pkg_resources import resource_filename
 
-from bio_hansel import program_name
-from bio_hansel.subtyping_params import SubtypingParams
+from biohansel import program_name
+from biohansel.subtype.subtyping_params import SubtypingParams
 
-SCHEME_FASTAS = {'heidelberg': {'file': resource_filename(program_name, 'data/heidelberg/tiles.fasta'),
+SCHEME_FASTAS = {'heidelberg': {'file': resource_filename(program_name, 'subtype/data/heidelberg/tiles.fasta'),
                                 'version': '0.5.0',
-                                'subtyping_params': SubtypingParams(low_coverage_depth_freq=20)},
-                 'enteritidis': {'file': resource_filename(program_name, 'data/enteritidis/tiles.fasta'),
+                                'subtyping_params': SubtypingParams(low_coverage_threshold=20)},
+                 'enteritidis': {'file': resource_filename(program_name, 'subtype/data/enteritidis/tiles.fasta'),
                                  'version': '0.7.0',
-                                 'subtyping_params': SubtypingParams(low_coverage_depth_freq=50)}}
+                                 'subtyping_params': SubtypingParams(low_coverage_threshold=50)}}
 
 COLUMNS_TO_REMOVE = '''
 pident
@@ -61,8 +60,5 @@ coverage
 qc_status
 qc_message
 """.strip().split('\n')
-
-REGEX_FASTQ = re.compile(r'^(.+)\.(fastq|fq|fastqsanger)(\.gz)?$')
-REGEX_FASTA = re.compile(r'^.+\.(fasta|fa|fna|fas)(\.gz)?$')
 
 JSON_EXT_TMPL = '{}.json'
