+++ biohansel/subtype/data/enteritidis/tiles.fasta
