+++ biohansel/utils.py
@@ -1,14 +1,15 @@
 # -*- coding: utf-8 -*-
 
 import logging
-from typing import List, Any, Optional, Tuple, Union
+from typing import List, Any, Tuple, Union
 
 import os
 import re
 from collections import defaultdict
 
-from .const import SCHEME_FASTAS, REGEX_FASTQ, REGEX_FASTA
-from .subtyping_params import SubtypingParams
+NT_SUB = {x: y for x, y in zip('acgtrymkswhbvdnxACGTRYMKSWHBVDNX', 'tgcayrkmswdvbhnxTGCAYRKMSWDVBHNX')}
+REGEX_FASTQ = re.compile(r'^(.+)\.(fastq|fq|fastqsanger)(\.gz)?$')
+REGEX_FASTA = re.compile(r'^.+\.(fasta|fa|fna|fas)(\.gz)?$')
 
 
 def does_file_exist(filepath: str, force: bool):
@@ -76,28 +77,6 @@ def find_inconsistent_subtypes(subtypes: List[List[int]]) -> List[str]:
     return incon_subtypes
 
 
-def get_scheme_fasta(scheme: str) -> str:
-    if scheme in SCHEME_FASTAS:
-        scheme_fasta = SCHEME_FASTAS[scheme]['file']
-    elif os.path.exists(scheme) and os.path.isfile(scheme):
-        scheme_fasta = scheme
-    else:
-        raise FileNotFoundError('Could not find user-specified subtyping scheme fasta "%s"', scheme)
-    return scheme_fasta
-
-
-def get_scheme_params(scheme: str) -> Optional[SubtypingParams]:
-    if scheme in SCHEME_FASTAS:
-        return SCHEME_FASTAS[scheme]['subtyping_params']
-
-
-def get_scheme_version(scheme: str) -> Optional[str]:
-    if scheme in SCHEME_FASTAS:
-        version = SCHEME_FASTAS[scheme]['version']  # type: str
-        return version
-    return None
-
-
 def collect_fastq_from_dir(input_directory: str) -> List[Union[str, Tuple[List[str], str]]]:
     fastqs = []
     for x in os.listdir(input_directory):
@@ -151,7 +130,6 @@ def collect_fasta_from_dir(input_directory: str) -> List[Tuple[str, str]]:
     return input_genomes
 
 
-NT_SUB = {x: y for x, y in zip('acgtrymkswhbvdnxACGTRYMKSWHBVDNX', 'tgcayrkmswdvbhnxTGCAYRKMSWDVBHNX')}
 
 
 def revcomp(s):
@@ -170,30 +148,74 @@ def is_gzipped(p: str) -> bool:
     return bool(re.match(r'^.+\.gz$', p))
 
 
-def init_subtyping_params(args: Optional[Any] = None,
-                          scheme: Optional[str] = None) -> SubtypingParams:
-    """Initialize subtyping parameters based on command-line arguments and scheme defaults
+def init_console_logger(logging_verbosity=3):
+    logging_levels = [logging.ERROR, logging.WARN, logging.INFO, logging.DEBUG]
+    if logging_verbosity > (len(logging_levels) - 1):
+        logging_verbosity = 3
+    lvl = logging_levels[logging_verbosity]
+
+    logging.basicConfig(format=LOG_FORMAT, level=lvl)
+
+
+def collect_inputs(files,
+                   input_fasta_genome_name,
+                   input_directory,
+                   paired_reads,
+                   **kwargs) -> Tuple[List[Tuple[str, str]], List[Tuple[List[str], str]]]:
+    """Collect all input files for analysis
+
+    Sample names are derived from the base filename with no extensions.
+    Sequencing reads are paired if they share a common filename name without "_\d".
+    Filepaths for contigs and reads files are collected from an input directory if provided.
 
     Args:
         args: ArgumentParser.parse_args() output
-        scheme: Scheme name e.g. "heidelberg"
 
     Returns:
-        SubtypingParams with user-supplied values then scheme defaults then global defaults loaded
+        List of (contig filename, sample name)
+        List of ([reads filepaths], sample name)
     """
-    subtyping_params = get_scheme_params(scheme)
-    if subtyping_params is None:
-        subtyping_params = SubtypingParams()
-    if args is not None:
-        if args.low_cov_depth_freq:
-            subtyping_params.low_coverage_depth_freq = args.low_cov_depth_freq
-        if args.max_missing_tiles:
-            subtyping_params.max_perc_missing_tiles = args.max_missing_tiles
-        if args.min_ambiguous_tiles:
-            subtyping_params.min_ambiguous_tiles = args.min_ambiguous_tiles
-        if args.max_intermediate_tiles:
-            subtyping_params.max_perc_intermediate_tiles = args.max_intermediate_tiles
-        if args.low_cov_warning:
-            subtyping_params.min_coverage_warning = args.low_cov_warning
+    input_genomes = []
+    reads = []
+    if files:
+        fastas = [x for x in files if REGEX_FASTA.match(x)]
+        fastqs = [x for x in files if REGEX_FASTQ.match(x)]
+        if len(fastas) > 0:
+            logging.info('# of input fastas %s', len(fastas))
+            for fasta_path in fastas:
+                fasta_path = os.path.abspath(fasta_path)
+                if os.path.exists(fasta_path):
+                    genome_name = genome_name_from_fasta_path(fasta_path)
+                    input_genomes.append((fasta_path, genome_name))
+                else:
+                    logging.error('Input fasta "%s" does not exist!', fasta_path)
+        if len(fastqs) > 0:
+            logging.info('# of input fastqs %s', len(fastqs))
+            grouped_fastqs = group_fastqs(fastqs)
+            logging.info('Grouped %s fastqs into %s groups',
+                         len(fastqs),
+                         len(grouped_fastqs))
+            reads += grouped_fastqs
+    if input_fasta_genome_name:
+        for fasta_path, genome_name in input_fasta_genome_name:
+            input_genomes.append((os.path.abspath(fasta_path), genome_name))
+    if input_directory:
+        logging.info('Searching dir "%s" for FASTA files', input_directory)
+        input_genomes += collect_fasta_from_dir(input_directory)
+        logging.info('Searching dir "%s" for FASTQ files', input_directory)
+        reads += collect_fastq_from_dir(input_directory)
+    if paired_reads:
+        for x in paired_reads:
+            if not isinstance(x, (list, tuple)):
+                logging.warning('Paired end reads not list or tuple %s', x)
+                continue
+            filenames = [os.path.basename(y) for y in x]
+            common_prefix = os.path.commonprefix(filenames)
+            genome_name = re.sub(r'[\W\_]+$', r'', common_prefix)
+            if genome_name == '':
+                genome_name = filenames[0]
+            reads.append((x, genome_name))
+    return input_genomes, reads
 
-    return subtyping_params
+
+LOG_FORMAT = '%(asctime)s %(levelname)s: %(message)s [in %(pathname)s:%(lineno)d]'
